odyssey-intro-strawberry/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
fastapi==0.115.12
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mypy==1.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
strawberry-graphql==0.270.4
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
